requirements.txt
asyncio
aiogram
sqlalchemy
python-dotenv
psycopg2